requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.33
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.21.0
Pillow>=9.0.0
Flask>=2.3.0
threading
time
logging
datetime
json
os
sys